requirements.txt
# Core scientific stack
numpy>=1.24
scipy>=1.10
pandas>=2.0
xarray>=2023.7.0
dask>=2023.5.0
netCDF4>=1.6
cfgrib>=0.9.14

# Deep learning
torch>=2.2
pytorch-lightning>=2.2

# IO / utils
requests>=2.31
PyYAML>=6.0

# Web framework
Flask>=3.0
flask-cors>=4.0
